Extract Trilogy::DatabaseStatements
To be consistent with other adapters.
